move disposal locations to new singular file
