+++ api/src/opentrons/protocol_api/__init__.py
@@ -24,8 +24,7 @@ from .module_contexts import (
 )
 from ._liquid import Liquid
 from ._types import OFF_DECK
-from ._trash_bin import TrashBin
-from ._waste_chute import WasteChute
+from ._disposal_locations import TrashBin, WasteChute
 from ._nozzle_layout import (
     COLUMN,
     ALL,
