+++ api/src/opentrons/protocol_api/core/engine/instrument.py
@@ -38,8 +38,7 @@ from . import deck_conflict
 from ..instrument import AbstractInstrument
 from .well import WellCore
 
-from ..._trash_bin import TrashBin
-from ..._waste_chute import WasteChute
+from ..._disposal_locations import TrashBin, WasteChute
 
 if TYPE_CHECKING:
     from .protocol import ProtocolCore
