+++ api/src/opentrons/protocol_api/create_protocol_context.py
@@ -22,7 +22,7 @@ from opentrons.protocols.api_support.definitions import MAX_SUPPORTED_VERSION
 
 from .protocol_context import ProtocolContext
 from .deck import Deck
-from ._trash_bin import TrashBin
+from ._disposal_locations import TrashBin
 
 from .core.common import ProtocolCore as AbstractProtocolCore
 from .core.legacy.deck import Deck as LegacyDeck
